requirements.txt

annotated-types
dataclasses-json
faiss
joblib
langchain
langchain-community
langchain-core
langchain-text-splitters
langsmith
langgraph
llama-index
llama-parse
multidict
mypy-extensions
numpy
pdfplumber
pyasn1_modules
pymupdf
pyparsing
python-dotenv
requests
scikit-learn
sentence-transformers
tokenizers
torch
tqdm=
transformers
